fixed the compilation errors in the test drivers

+++ elina_oct/tests/libFuzzer/test_oct0.c
@@ -8,7 +8,7 @@
 
 extern int LLVMFuzzerTestOneInput(const uint64_t *data, size_t dataSize) {
 	unsigned int dataIndex = 0;
-	unsigned short int dim = MIN_DIM;
+	size_t dim = MIN_DIM;
 	if (make_fuzzable_dimension(&dim, data, dataSize, &dataIndex)) {
 		elina_manager_t * man = opt_oct_manager_alloc();
 		opt_oct_t * top = opt_oct_top(man, dim, 0);
